Add comformance to ImageService for PHImageManager
